KFileItem: handle KFileSystemType::FuseBlk in isSlow()
With the change in KFileSystemType, "fuseblk" won't be detected as KFS::Nfs
type, but it still could be a potentially slow filesystem, so mark it as
such.

KFileSystemType will report the type as FuseBlk_BlockDevice for local disk
partitions mounted as "fuseblk", this means that e.g. NTFS filesystems
mounted as "fuseblk" are not marked as slow.

See KFileSystemType API docs (and the previous commit in KMountPoint) for
more details.