Fix missing details in message dialogs from worker
The port to AskUserActionInterface in commit
5f7fe0501725564d4c8247f74731bcee52cd711d missed out to support the hack
introduced to allow workers passing details for message box types of
WarningContinueCancelDetailed (initially introduced for privilege
confirmation dialogs).

While JobUiDelegateExtension had no argument for details in its
requestMessageBox() API and the same hack as used with the SlaveBase
messageBox() API was done, passing details via metadata in a special
entry called "privilege_conf_details", AskUserActionInterface though
has a details argument in its requestUserMessageBox() call, so
SlaveInterface now just has to cache the string and pass it on
with the next request for a messagebox of type WarningContinueCancelDetailed.

Also updating the TODOs related to the hack with a consistent id pointing
to a central place explaining the hack and how to improve things.

(cherry picked from commit 6198e59fa2e3d4f71ac12fd7d616757c207fb407)