Warning dialogs: Change button text from "Delete" to "Delete Permanently"
In dialogs where the user is asked whether they want to permanently delete
something, It's possible a user might ignore the warning and only focus on
the buttons. If they only see "Delete," they might click it thinking it
will simply send the files to the trash, and mistakenly delete their files
permanently without realizing it.

This commit changes the button text from "Delete" to "Delete Permanently"
so the user has another chance to read what they're about to do and avoid
performing a destructive action.