Ignore finished() from the timeoutSpecialCommand
If a timeoutSpecialCommand is set, SlaveBase triggers a call to the Slave's
special() method after the given timeout expires. This call is triggered as
if it was coming from the application, so it must be taken care of that it
does not actually send any result, which the application might not expect.
In particular in the case of FileJob, calling finished() stopped the Job
immediately, the result() signal is emitted with no error set.

Ignore calls to finished() and error() from special() in that case.

BUG: 459980