Remove docs for telnet scheme-handler
In khelpcenter only shown in "Application Manuals/Lost & Found" section,
as the installation to kioslave5 subdir does not result in the potentially
expected listing in the "KIO workers" section (which also would not be
that correct).

Also does the current text not really provide anything useful, and also
misses to document the other, these days more important schemes like
ssh that the ktelnetservice executable supports.