Remove default Search items from Places panel
These items take up space in the Places Panel and make its contents very
long and harder to visually parse because of all the items. They are
just not useful enough for the visual cost they impose, so let's not
add them to the Places panels of new installs going forward. Existing
installs will be unaffected, and the functionality can still be
accessed using the standard Find feature in Dolphin or even re-added
manually to the Places panel by people who do find them useful.