when inside a sandbox use different openwith behavior
inside a sandbox there are no useful services for us to enumerate so we
can only offer a generic open with action. do that and call openURI when
that action is triggered for all items in the collection. this is a bit
awkward but the best we can do for now.