Pre-select "Delete" in Delete Confirmation Dialog
This commit makes the "Delete" button the default button in the dialog
for confirming direct deletion of a file and in the dialog for
confirming emptying of the trash.

This was accidentally changed in
8144733c5110c66922381f97829f9470b26c8122 while trying to change the
dialog styling to be more like a warning. Users were unhappy about this
because it broke their muscle memory of "Shift+Del" into "Return" to
permanently delete a file.

This "revert" changes the dialog type from "warning" back to
"question". However, this would also change the icon displayed in the
dialog so the icon is explicitly set to be the same as in the "warning"
dialog.

An alternative suggestion was to add explicit API to select the
pre-selected button but this didn't happen as part of this "revert".

BUG: 462845