Convert desktop files of KCMs to json
Now we can get rid of the conversion logic without duplicating the files.
The kcmutils cmake function is a mock for now, because kcmutils depends on kdeclarative which depends on KIO.
But the KDeclarative dependency will be removed in Kcmutils and we plan on moving the KCMs to kio-extras later on.