Port KIO::Core away from QTextCodec and thus Qt6::Core5Compat
KRemoteEncoding::encodingMib cannot be implemented on top of
QStringConverter, but is fortunately unused.

CHECK_UTF8 was never defined, so that part has been dead code.

The local codec is always assumed to be UTF-8, so we can just as well
hardcode that in SessionData. Thus we also assume UTF-8 is always
available and doesn't need a fallback to Latin1.