Remove LegacyCodec
This used to override the built-in UTF-8 codec (!) with an alternative
version that was more fault-tolerant to mis-encoded local file names.
This however has no effect now that QTextCodec isn't used by QFile
anymore, and there is no such mechanism for an external codec override
anymore.

Should the original problem still exist, this would need to be addressed
inside Qt.