ApplicationLauncherJob: find mimetype before showing open with dialog
To support "last used application" feature in Plasma Integration, it's
required to pass a mimetype name in `promptUserForApplication`.
This also matches the behavior in `KIO::OpenUrlJob`.

Implemented in KF5 because it's unlikely to break things and can support Qt5
applications to use the "last used application" feature" after Plasma 6
is released.