kprocessrunner: Replace env based launchMode lookup with runtime detection
The env guard was introduced as we wanted to do a soft launch and not
risk regressions to users, only changing behaviour when a newer Plasma
was used.

The rollout has been a non-issue, so we can just opt-in directly without
needing to rely on Plasma setting things. This will also improve cgroup
compatibility when run under Gnome.

The env values are left as they're used by the unit test.