delete threaded workers on a thread
this prevents potential deadlocks between the worker thread and the main
thread in case the former relies on the latter to do work. specifically
qdbus seems unhappy if we wait in the main thread. in point of fact we
shouldn't be waiting in the main thread anyway since the result of the
thread is immaterial

(wait is also called inside the workerthread destructor, I'm leaving it
there as a safe guard against deleting still-running threads)

BUG: 468673