WidgetsAskUserActionHandler: make sure all dialogs are created in the main thread
In Qt, the GUI runs in its own thread and must be accessed from the
main thread. If QWidget is created in a non-main thread, we can use
the QMetaObject::invokeMethod() method to create it in the main thread.

BUG: 458709
CCBUG: 364039
FIXED-IN: 5.106
(cherry picked from commit ac2a594a654a3821e94b34be92d534608b9837a1)