Remove automatic cookie persistence in the HTTP worker
This is a leftover from times when this was mainly used for a web browser.
For the remaining uses of this automatic cookie storage is generally not
desired though, and therefore disabled explicitly already. Should
applications want to persist cookies, that is still possible, but needs
to be done explicitly now, same as with QNAM.

The bigger goal here is to eventually move the KCookieJar kded module and
KCM to its only remaining users, Konqueror.