KUrlRequester: restore unnamed filter compatibility with Plasma file dialog
6e945ed1dc01ffbaf30455341cff0ba945877255 changed the behaviour of
KUrlRequester::setFilter to not generate a name for the Qt-style filter
syntax variant if also none was given to the KDE-style variant.

The missing name in the filter though triggers a bug with Plasma's
platform QFileDialog implementation, resulting in the drop of the filter
entry.

So to restore the compatibility and matching API user expectations, the
filters, when passed to the actual file dialog instance, are preprocessed
to inject any missing explicit name. As before the name used is the
one of the filter expressions.