Rewrite HTTP worker based on QtNetwork classes
KIO's HTTP implementation is massive, old, and heavily influenced by requirements from Konqueror.

The size and complexity of the code make it hard to maintain.

With Konqueror not using KIO HTTP any more (nor any other browser) a lof of the features are not
relevant any more.

We can leverage QNetworkAccessManager to do most of the heavy HTTP lifting and only do the KIO
integration on top as well as webdav support

This drops support for a number of metadata keys that are unused in KDE code.

It comes with a new set of autotests that runs various jobs against a local HTTP server

It also introduces a new way of handling SSL errors. Instead of doing it via the messagebox
worker infrastructure it introduces dedicated worker commands for it, which is a lot cleaner