Avoid having 4 webshortcut reloads when using kurifilters
We load the entries once the registry is instantiated, but also in case
the config has changed. Using a boolean we can skip this in case it is
called from the constructor.

The other major issue was that the Q_GLOBAL_STATIC was in multiple
compilation units. Thus we did the duplicate reparsing in two places.
By having a shared lib, we use the same Q_GLOBAL_STATIC in both plugins.