KProtocolInfoFactory: Allow to set bool flag to not reload protocol cache
In case of webshortcuts and other code being used in KRunner plugins, we would trigger the cache reloading quite often.
Because the checks are like "is this webshortcut a protocol known to KIO?", it is not worth reloading the cache.

In case one types an unknown protocol in dolphin, it is reasonable to
reload the cache. That is why the default param is set to true.