KCoreDirListerCache:: make sure not to reinsert an item twice
In slotUpdateResult and processPendingUpdates.

pendingUpdates can be updated in KCoreDirListerCache::slotFilesChanged
called synchrounously when file on disk are changed or asynchrously when
its parent directory is scanned for whatever reason. It can happen that
the pendingUpdateTimer triggers as the file is updated, this may cause
reinsert for the same url, the second one causing a crash.

Not 100% sure this is sufficient to fix the crash, so only CC.

CCBUG: 473488