Fix loading desktopfile-based service actions
The code before 0eb24754b5e3fb0025889bca1fca826f45961254 would check

if (isBuiltin || !serviceAction.noDisplay())

with the latter part getting lost in the change. This results in olny builtin actions getting picked.
Since the assumption seems to be that actions are never noDisplay we can drop the entire check

BUG: 477315