KDirModel: Consider invalid roots are local fs
If the item set on a node (typically the root node) was null it was
implicitly considered local fs. But if the item was not null and the
url invalid, it'd be considered network fs.

In the context of KDirModel caching of the fs type guessing this would
mean we'd wrongly assume the whole tree is network fs even if it's not.

That's why we change this to consider non-null items with invalid URLs
as local fs.

BUG: 477039