[kprocessrunner] Clean up maybeAliasedName
maybeAliasedName does two things:

- resolving the potential alias for m_service
- Fill in a given pattern with the result and a UUID

Mixing those two things in a single function is dangerous.

For requestXdgActvationToken we want to resolve the alias, but not
do any of the pattern/UUID stuff that is only for the system runners.

Shuffle things around to make it more clear.

Also make sure that m_service is set before resolveServiceAlias is called

BUG: 478835