KFileItem: Use stat data for isDir if we already performed a stat
`KFileItem::isDir()` was changed to use the mime type if it is available
to avoid doing a stat call. However, if we already have that data
available, using the stat information is much faster than doing a mime
type lookup. So first check to see if we already have that information
and if so, use it.