connectionbackend: improve jumbo packet handling
(also should fix things on windows)

jumbo handling was a bit weird in that we kept bumping the buffer size
sky high. and then eventually (hopefully) would be able to read
everything in one go and emit. In practice that seems to have not worked
out on windows. do partial reads instead.

to facilitate this change the existing Task struct is now serving as
container we read into so it becomes a bit easier to manage the state.

this should also make the code a bit easier to reason about. when we
have no task we now start a new task. if we have a task we always read
as much data as possible into the task. when we read all data then the
task gets emitted.

BUG: 481311