kioexec: Remove temporary files after the main loop has completed
It is possible to launch kioexec with multiple file or URL arguments,
if the receiving application accepts them.  If these are temporary files
(indicated by the '--tempfiles' option), they were deleted one after the
other in slotRunApp() loop with a 180 second pause between each one.
This means that the process remained active for a very long time.

Just accumulate the names of temporary files to be deleted inside the
loop, and then sleep for that time just once.  Remove all of the files
together after the sleep is complete.