KFilePlacesItem: Show teardown busy indicator during optical media eject
Ejecting an optical disc (e.g. DVD) can take some time, show the busy
indicator while this is in progress. It already shows the busy indicator
while mounting it (through Storage Access) but was lacking the indicator
when ejecting it. This also makes it consistent with when unmounting
a USB stick.