knewfilemenu: add isNewDirNameJobRunning
As per discussion here: https://invent.kde.org/system/dolphin/-/merge_requests/747#note_903955

In order to avoid spawning multiple popups for network folders when creating new 
directories with spamming the `Ctrl+Shift+N` shortcut in Dolphin,
we need to know if the nameJob that spawns those popups is still running.

This adds a simple bool method that returns if the nameJob is still being ran.
In Dolphin this can be used to check if there's already job for spawning this popup
running, and if there is, we can ignore the shortcut.