Properties dialogue URL tab: Fix vertical alignment
The second parameter to QBoxLayout::addWidget() when adding the
label 'l' is a stretch factor, not an alignment.  Setting this
(with a value of Qt::AlignRight = 2) applied the largest stretch
factor to the label rather than the bottom stretch added later,
so pushing the URL text box down to the middle of the dialogue.