KAbstractFileItemActionPlugin: Remove wrong sentence in documentation
The removed line says to have the KAbstractFileItemActionPlugin as
the parent of the returned actions. Contrarily, a few lines further
down the documentation says to have the parentWidget as the parent
of the returned actions.

The latter is the more recent addition to the documentation which
was added because the former did not work out in some contexts.

The bug mentioned below contains a list of changes that made the
parentWidget the parent of actions from
KAbstractFileItemActionPlugins, which seems to have worked out
because from what I can tell they were not reverted in the years
since then.

CCBUG: 425997