Use resolved executable in KProcessRunner::fromExecutable
The other two code paths do this implicitely via KProcess::setShellCommand
and DesktopExecParser::resultingArguments. This does not matter
when the current process is starting the new process but is needed
when the new process is going to be started by systemd because it
might have a different PATH than the current process.
A new test is added that tests the case of a binary only being
in the PATH of the current process.