previewjob: Port to QImageReader
When we want to generate a thumbnail we load metadata from the existing
cached image to work out if we want to return the current cached image
or create a new one.

At that point in time we don't need to load the entire image as it could
end up being thrown away.