dropjob.cpp: Return if itemProps has no valid urls
Check that url is valid before appending it to fileItems.
After that, make sure that itemProps.items is not empty.
If it's empty, return and do nothing. Otherwise plasmashell will crash,
when trying to drag and drop an item from task manager that has
no associated desktop file, for example.

BUG:484674