Remove parent for DropMenu
QWidget::winId() will force creating window handles for all its ancestor
widgets unless the Qt::WA_DontCreateNativeAncestors flag is set.

For example, on wayland, this would result in creating sub-surfaces, and
depending on the client, it's likely that there are going to be issues with
painting.

On the other hand, since the DropJob takes care of setting the transient
parent for the DropMenu, the parent can be simply omitted.

BUG: 490183