Lookup QWidget parent at runtime
In case the QWidget has been associated with the drop job while not
being fully integrated into the widget tree, KJobWindows::window()
would return null instead of a proper transient parent.

Changing the api of KJobWidgets::setWindow() is also challenging. So
proceed with querying the window handle of the widget at runtime.