dropJob: special-case "downloading http URLs" drop with better text
When dropping URLs, the "Copy" action now has its text and icon
changed to reflect that this is really more of a download job than a
copy job. Technically it's a copy, but the more natural way and common
way to express this kind of thing is to say that it's a download.