knewfilemenu: Make ~/Templates work by simply placing files and folders there
Adds functionality so that placing any file in `~/Templates` creates the same file in the context menu as a template. This also works for directories so you are able to create project skeletons. I think I accidentally made it so that if you already have a `.desktop` file in `~/.local/share/templates` that points to `~/Templates`, the `.desktop` file will take priority and there is no duplication. 

BUG: 191632