KFilePlacesItem: Use Solid to find home mount point
KMountPoints potentially parses MTAB which can be slow.
Instead, use Solid which already has a cache of all the devices.

The impact on a typical system is negligible but real-world
testing on a system with lots of network shares showed a
speed-up of over 80% in this code path.

Signed-off-by: Kiriakos Antoniadis <kiriakos.antoniadis@advantest.com>

(cherry picked from commit b5d45a857e6fdb3fb7b08e44753eef203ea605c6)