KFilePlacesView: have setUrl() handle trailing slashes in place URLs
Whilst KFilePlacesView::setUrl() correctly strips its argument when
comparing it to place URLs, it fails to strip the place URL itself. This
means that if a place URL contains a trailing slash, it is not correctly
highlighted in the place view.

Instead of comparing both QUrl objects directly, use QUrl::matches,
which allows us to pass QUrl::StripTrailingSlash to both.

Add a test case for this behaviour as well.