previewjob: Use .cache as temp folder, delete temp file after use
Current solution moves items into `/tmp/` which is not good, since
any user can see thumbnail images from there. The files
are also not deleted after the image is generated.

This moves the tmp files into `~/.cache/thumbnails/size/tmp/` and
then when the thumbnailer is done, the file is removed.

We also do not use standard thumbnailers if caching is turned off, since we
do not have thumbnail folder in the first place, and the thumbnailers rely on that.

BUG:494096
CCBUG:494061