Clean up Properties dialog to follow HIG, improve UX, remove frames and fix padding regression
Plasma 6 introduced a regression in this dialog where there are no longer proper margins in most tabs, due to the move to a frameless style. This makes specifically URL and Application properties tabs look quite bad, as they end up with text glyphs right up against the edge of the window. Explicit 0px margins are removed, instead leaving the layouts to inherit from the stylesheet, which also improves maintainability.

The dialog can no longer be shrunk to the point where the tab contents look broken.

In the Permissions pane, the QGroupBox is now flat, so it looks the same as Kirigami.FormLayout, and doesn't have unnecessary frames. It's also ported to QFormLayout.

QLabel in KChecksumsPlugin is now a readonly QLineEdit, as this shows a nice looking box, and doesn't unexpectedly expand the window by an insane amount, where it can't be resized back down to a normal size.

BUG: 484789