Check that admin worker was installed by root
The worker behind the "admin" protocol typically asks for elevated
permissions as soon as it is started. Therefore, if malware hides
behind that protocol it will easily get full system access.

Plugins like the KIO admin worker are searched for in various paths
and those search paths can even be modified by applications
( https://doc.qt.io/qt-6/qt-conf.html ), so it seems like we can
not guarantee that the plugin would have been legitimately
installed prior to this commit. Prior to this commit, the plugin file
simply has to sit in a folder that KIO searches for. So if a user
"installed" (without entering a password) an application in user space
which uses KIO and that application lists a random folder as a library
search path (which is possible by providing a qt.conf file as the
documentation https://doc.qt.io/qt-6/qt-conf.html explains), then any
file put there can hide behind the "admin" protocol name and KIO will
happily load it.

That's just one example. The general point is that all of this seems
like quite a big attack surface. It just needs one of the folders that
Qt searches in to be set up badly and any executable/script run as user
can move malware to that location that KIO will load. Users would never
expect that. The check added in this commit seems like a harmless way
to bring security to a multi-layered loading system that every
distribution can mess with by providing a qt.conf file.

This commit checks that the plugin is owned by root and that other
groups and users can not modify it to make sure that an administrator
has already actively put that specific software on the system for this
specific usage.

This security requirement is only ignored when KIO itself also already
fails this security check. This way we do not hinder development
environments from starting kio-admin. The security is already
compromised under those conditions anyway, so we are not gaining
anything by blocking execution of kio-admin there.