ApplicationLauncherJob: don't launch with a broken symlink argument
If there is only one url argument, it's a local path, and it's a broken
symlink (or not exist in general), don't launch the application and fail
immediately.

If there is multi url arguments, and one of them is a broken symlink,
the application will be launched with all urls including the broken one.