core/copyjob: report speed when copying multiple files
The speed measurement is done in WorkerInterface::calcSpeed currently.
Except it can only measurement for a single file operation.

Add a speed measurement to CopyJob so copying speed is properly reported
even when copying multiple files, each taking less than a second to
copy.

The speed is smoothed accross the last 8 seconds.

CCBUG: 391199