[webdav] Handle redirections which add trailing slashes
KIO doesn't always send trailing slashes when referencing collections.
The RFC (quoted in the comment) requires that such clients handle
redirections, which KIO can't do correctly in this case.

Add explicit handling for such redirections in the http worker, such that
KIO and applications don't need to be aware of this.

CCBUG: 484580