[KFilePlacesView] Improve automatic resize heuristic
The old logic always chose the smallest possible icon size if any of
the items in the places view had a name that is too long to fit into
the view. This meant that in most typical scenarios, in which a user
has at least one device with a long name, the smallest possible icon
size was chosen.

This commit makes it so the widest 20 percent of texts are ignored
if there is a big difference in width between them and the widest
of the other 80 percent. This way, bigger icon sizes are chosen if
everything fits aside from the text of these few unusually wide names.

The basic advantage of this change is that users have way taller
click targets for the items on the places panel when there is enough
space. If users do not like this, they were never supposed to use the
automatic resize behaviour to begin with, and they can set a fixed
size.

BUG: 449544