KDirOperator: improve handling of forward/back mouse buttons
Since the the forward and back mouse buttons would
be passed through to the underlying Qt widget which handles it
like any other click, this meant single-clicking with these
buttons would allow selecting and dragging items, and double-clicking
would open files or folders, instead of just navigating through the
history as expected.

These buttons are now handled when pressed, not released; this
preventing being able to drag items, and bringing behaviour in
line with Dolphin. Double-click events for these buttons are now
properly handled. The events are then discarded to stop single-clicks
selecting items, or double-clicks opening files or folders.

CCBUG: 443169