KUrlNavigatorDropDownButton: Add text and tooltip
Prior to this commit the purpose of this button was only
communicated by its position at the left end of the path and its
non-descriptive arrow icon.

This commit adds a text and tooltip to the button. This is
especially crucial for accessibility because until now the button
had zero accessible information about its purpose because neither
its position nor its icon were accessible e.g. to blind users.