KUrlNavigator: Fix Tab order
KUrlNavigator contains a lot of widgets which are dynamically shown
or hidden under various conditions. This messes with the keyboard
focus order which can lead to inane focus orders like losing focus
to widgets outside of KUrlNavigator instead of going to the
neighboring button.

This commit enforces a sensible Tab key focus order. For
left-to-right languages, focus goes left-to-right through all
visible child widgets. The leftmost such widget is set as the
focusProxy() of this KUrlNavigator.

For right-to-left languages, the above logic is mirrored.