widgets/clipboardupdater: send mimedata with uuid so clipboard history can replace the old clip
ClipboardUpdater tries to update the current clipboard to replace moved
file entries with new urls. However, this can mess the clipboard history
when file operations are slow. This sends an uuid in the updated
mimedata so clipboard history can read it and replace the old clip based
on the uuid.