DropJob: Track supportedActions in dropEvent before showing menu
The source of a drag will provide possible actions allowed for the drag.
One action will then be selected by layers below KIO.

When performing a drop on a folder KIO will prompt the user as to which
action should be performed. This fails to take into account what the
source allows.