dropjob: Check if file is already in trash during dropjob
Previously we checked if file is already in trash during the dropjob,
but only by using the scheme. We should also check using the xdg path,
to make sure we are not moving the same file to same folder, causing it
to rename and confusing users.

This also adds unittest `shouldDropFromXdgTrashToTrash`.

BUG:497390